Rename Contract component to Contact
